split to region screen
