+++ src/client/screens/Westeros/index.jsx
@@ -1,42 +1,13 @@
 import React, {PropTypes, Component} from 'react';
 import Relay from 'react-relay';
 import Region from './components/Region';
-import List from 'components/List.jsx';
-import Modal from 'components/Modal.jsx';
+
 import './index.less';
 
 export class Westeros extends Component {
   static propTypes = {
     viewer: PropTypes.object
   };
-  static childContextTypes = {
-    showModal: PropTypes.func
-  };
-
-  state = {
-    showModal: false,
-    modalContent: null
-  };
-
-  hideModal = () => {
-    this.setState({
-      showModal: false,
-      modalContent: null
-    });
-  };
-
-  showModal = (content) => {
-    this.setState({
-      showModal: true,
-      modalContent: content
-    });
-  };
-
-  getChildContext() {
-    return {
-      showModal: this.showModal
-    };
-  }
 
   getRegion(region, index) {
     return <Region key={index} region={region}/>;
@@ -52,8 +23,6 @@ export class Westeros extends Component {
     );
   }
 }
-
-// <Modal onRequestClose={this.hideModal} isOpen={this.state.showModal}>{this.state.modalContent}</Modal>
 
 export default Relay.createContainer(Westeros, {
   fragments: {
