+++ src/client/shared/core/routes.jsx
@@ -26,11 +26,11 @@
 import React from 'react';
 import { Route } from 'react-router';
 import Westeros from '../../screens/Westeros';
-import Family from '../../screens/Westeros/screens/Family';
+import Region from '../../screens/Westeros/screens/Region';
 import queries from './queries';
 
 export default (
   <Route path="/" component={Westeros} queries={queries}>
-    <Route path="family" component={Family} queryParams={['regionId']} queries={queries} />
+    <Route path="region" component={Region} queryParams={['regionId']} queries={queries} />
   </Route>
 );
